rename notificationstore to notifications and put in state lib
