+++ components/outcomes/ScalarReportBox.tsx
@@ -9,7 +9,7 @@ import TransactionButton from "components/ui/TransactionButton";
 import Decimal from "decimal.js";
 import { ZTG } from "lib/constants";
 import { useSdkv2 } from "lib/hooks/useSdkv2";
-import { useNotifications } from "lib/stores/NotificationStore";
+import { useNotifications } from "lib/state/notifications";
 import { useStore } from "lib/stores/Store";
 import { getCurrentPrediction } from "lib/util/assets";
 import { extrinsicCallback, signAndSend } from "lib/util/tx";
