+++ lib/stores/Store.ts
@@ -19,7 +19,7 @@ import { isAsset, ztgAsset } from "../types";
 import Wallets from "./wallets";
 import ExchangeStore from "./ExchangeStore";
 import MarketsStore from "./MarketsStore";
-import NavigationStore from "../state/notifications";
+import NavigationStore from "./NavigationStore";
 import PoolsStore from "./PoolsStore";
 
 import { Context, Sdk } from "@zeitgeistpm/sdk-next";
